cambios

+++ G06P02/src/main/FitnessComparatorMin.java
@@ -8,6 +8,6 @@ public class FitnessComparatorMin implements Comparator<Individuo> {
 	@Override
 	public int compare(Individuo o1, Individuo o2) {
 		
-		return Double.compare(o1.getFitness(), o2.getFitness());
+		return Integer.compare(o1.getFitness(), o2.getFitness());
 	}
 }
